Tweaked A LOT

+++ gamemode/items/base/sh_bags.lua
@@ -1,32 +1,23 @@
-local INVENTORY_TYPE_ID = "grid" -- The ID of the inventory type used for the bag
-ITEM.name = "Bag" -- The name of the item
-ITEM.desc = "A bag to hold more items." -- The description of the item
-ITEM.model = "models/props_c17/suitcase001a.mdl" -- The model of the item
-ITEM.category = "Storage" -- The category of the item
-ITEM.isBag = true -- Indicates that the item is a bag
+local INVENTORY_TYPE_ID = "grid"
+ITEM.name = "Bag"
+ITEM.desc = "A bag to hold more items."
+ITEM.model = "models/props_c17/suitcase001a.mdl"
+ITEM.category = "Storage"
+ITEM.isBag = true
+ITEM.BagSound = {"physics/cardboard/cardboard_box_impact_soft2.wav", 50}
+ITEM.invWidth = 2
+ITEM.invHeight = 2
 
--- Sound played when interacting with the bag
-local SOUND_BAG_RESPONSE = {"physics/cardboard/cardboard_box_impact_soft2.wav", 50}
-
-ITEM.invWidth = 2 -- The width of the bag's inventory grid
-ITEM.invHeight = 2 -- The height of the bag's inventory grid
-
--- The "View" function of the item
 ITEM.functions.View = {
-    icon = "icon16/briefcase.png", -- The icon used for the function
+    icon = "icon16/briefcase.png",
     onClick = function(item)
         local inventory = item:getInv()
         if not inventory then return false end
         local panel = lia.gui["inv" .. inventory:getID()]
         local parent = item.invID and lia.gui["inv" .. item.invID] or nil
-
-        if IsValid(panel) then
-            panel:Remove()
-        end
-
+        if IsValid(panel) then panel:Remove() end
         if inventory then
             local panel = lia.inventory.show(inventory, parent)
-
             if IsValid(panel) then
                 panel:ShowCloseButton(true)
                 panel:SetTitle(item:getName())
@@ -36,11 +27,9 @@ ITEM.functions.View = {
             local index = item:getData("id", "nil")
             ErrorNoHalt("Invalid inventory " .. index .. " for bag item " .. itemID .. "\n")
         end
-
         return false
-    end, -- The function to execute when the function is clicked
-    onCanRun = function(item) return not IsValid(item.entity) and item:getInv() end -- The function to determine if the function can be run
-    
+    end,
+    onCanRun = function(item) return not IsValid(item.entity) and item:getInv() end
 }
 
 function ITEM:onInstanced()
@@ -50,31 +39,31 @@ function ITEM:onInstanced()
         h = self.invHeight
     }
 
-    lia.inventory.instance(INVENTORY_TYPE_ID, data):next(function(inventory)
-        self:setData("id", inventory:getID())
-        hook.Run("SetupBagInventoryAccessRules", inventory)
-        inventory:sync()
-        self:resolveInvAwaiters(inventory)
-    end)
+    lia.inventory.instance(INVENTORY_TYPE_ID, data):next(
+        function(inventory)
+            self:setData("id", inventory:getID())
+            hook.Run("SetupBagInventoryAccessRules", inventory)
+            inventory:sync()
+            self:resolveInvAwaiters(inventory)
+        end
+    )
 end
 
 function ITEM:onRestored()
     local invID = self:getData("id")
-
     if invID then
-        lia.inventory.loadByID(invID):next(function(inventory)
-            hook.Run("SetupBagInventoryAccessRules", inventory)
-            self:resolveInvAwaiters(inventory)
-        end)
+        lia.inventory.loadByID(invID):next(
+            function(inventory)
+                hook.Run("SetupBagInventoryAccessRules", inventory)
+                self:resolveInvAwaiters(inventory)
+            end
+        )
     end
 end
 
 function ITEM:onRemoved()
     local invID = self:getData("id")
-
-    if invID then
-        lia.inventory.deleteByID(invID)
-    end
+    if invID then lia.inventory.deleteByID(invID) end
 end
 
 function ITEM:getInv()
@@ -83,15 +72,11 @@ end
 
 function ITEM:onSync(recipient)
     local inventory = self:getInv()
-
-    if inventory then
-        inventory:sync(recipient)
-    end
+    if inventory then inventory:sync(recipient) end
 end
 
 function ITEM.postHooks:drop()
     local invID = self:getData("id")
-
     if invID then
         net.Start("liaInventoryDelete")
         net.WriteType(invID)
@@ -103,26 +88,21 @@ function ITEM:onCombine(other)
     local client = self.player
     local invID = self:getInv() and self:getInv():getID() or nil
     if not invID then return end
-    -- If other item was combined onto this item, put it in the bag.
     local res = hook.Run("HandleItemTransferRequest", client, other:getID(), nil, nil, invID)
     if not res then return end
-
-    -- If an attempt was made, either report the error or make a
-    -- "success" sound.
-    res:next(function(res)
-        if not IsValid(client) then return end
-        if istable(res) and isstring(res.error) then return client:notifyLocalized(res.error) end
-        client:EmitSound(unpack(SOUND_BAG_RESPONSE))
-    end)
+    res:next(
+        function(res)
+            if not IsValid(client) then return end
+            if istable(res) and isstring(res.error) then return client:notifyLocalized(res.error) end
+            client:EmitSound(unpack(self.BagSound))
+        end
+    )
 end
 
 if SERVER then
     function ITEM:onDisposed()
         local inventory = self:getInv()
-
-        if inventory then
-            inventory:destroy()
-        end
+        if inventory then inventory:destroy() end
     end
 
     function ITEM:resolveInvAwaiters(inventory)
@@ -138,14 +118,12 @@ if SERVER then
     function ITEM:awaitInv()
         local d = deferred.new()
         local inventory = self:getInv()
-
         if inventory then
             d:resolve(inventory)
         else
             self.awaitingInv = self.awaitingInv or {}
             self.awaitingInv[#self.awaitingInv + 1] = d
         end
-
         return d
     end
 end
